Style refactor in views. MOve recovery logic to manage.
