+++ src/direct-auth-idx/direct-auth-idx/Views/Manage/SelectAuthenticator.cshtml
@@ -6,7 +6,7 @@
 }
 
 <hr />
-<div class="row">
+<div class="row d-center">
     <div class="col-md-8">
         <section id="selectAuthenticatorForm">
             @using (Html.BeginForm("SelectAuthenticatorAsync", "Manage", new { ReturnUrl = ViewBag.ReturnUrl }, FormMethod.Post, new { @class = "form-horizontal", role = "form" }))
@@ -26,33 +26,35 @@
                 @Html.ValidationSummary(true, "", new { @class = "text-danger" })
                 @Html.HiddenFor(m => m.PasswordId)
                 @Html.HiddenFor(m => m.PhoneId)
-                <ul>
-                @foreach (var authenticator in Model.Authenticators)
-                {
-                    <div>
-                        <label>
-                            @Html.RadioButtonFor(m => m.AuthenticatorId, authenticator.AuthenticatorId)
-                            @authenticator.Name
-                        </label>
-                    </div>
-                }
+                <ul class="no-padding-left">
+                    @foreach (var authenticator in Model.Authenticators)
+                    {
+                        <div>
+                            <label>
+                                @Html.RadioButtonFor(m => m.AuthenticatorId, authenticator.AuthenticatorId)
+                                @authenticator.Name
+                            </label>
+                        </div>
+                    }
                 </ul>
                 @Html.ValidationMessageFor(m => m.AuthenticatorId, "", new { @class = "text-danger" })
                 <div class="form-group">
                     <div class="col-md-offset-0 col-md-10">
-                        <input type="submit" value="Submit" class="btn btn-default" />
+                        <input type="submit" value="Submit" class="btn btn-primary btn-stretch-wide" />
                     </div>
                 </div>
-            }
-
-            @if (Model.CanSkip)
-            {
-                <hr />
-                using (Html.BeginForm("SkipAuthenticatorSelectionAsync", "Manage", new { ReturnUrl = ViewBag.ReturnUrl }, FormMethod.Post, new { @class = "form-horizontal", role = "form" }))
+                @if (Model.CanSkip)
                 {
-                    <input type="submit" name="skip" value="Skip this step" class="btn" />
-                }
+                    using (Html.BeginForm("SkipAuthenticatorSelectionAsync", "Manage", new { ReturnUrl = ViewBag.ReturnUrl }, FormMethod.Post, new { @class = "form-horizontal", role = "form" }))
+                    {
+                        <div class="form-group">
+                            <div class="col-md-offset-0 col-md-10">
+                            <input type="submit" name="skip" value="Skip this step" class="btn btn-info btn-stretch-wide" />
+                            </div>
+                        </div>
+                    }
 
+                }
             }
         </section>
     </div>
