+++ src/direct-auth-idx/direct-auth-idx/Views/Manage/SelectRecoveryAuthenticator.cshtml
@@ -11,7 +11,7 @@
     <div class="col-md-8">
         <section id="selectAuthenticatorForm">
 
-            @using (Html.BeginForm("SelectRecoveryAuthenticatorAsync", "Account", new { ReturnUrl = ViewBag.ReturnUrl }, FormMethod.Post, new { @class = "form-horizontal", role = "form" }))
+            @using (Html.BeginForm("SelectRecoveryAuthenticatorAsync", "Manage", new { ReturnUrl = ViewBag.ReturnUrl }, FormMethod.Post, new { @class = "form-horizontal", role = "form" }))
             {
                 @Html.AntiForgeryToken()
                 <h4>Reset your password</h4>
